backend/requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
statsmodels>=0.13.0
requests>=2.25.0
tqdm>=4.62.0
pyyaml>=6.0
plotly>=5.0.0

# API server
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30
sse-starlette>=2.1
pydantic>=2.7
python-multipart>=0.0.9
